+++ packages/cosmwasm-stargate/src/index.ts
@@ -1,3 +1,7 @@
 export { cosmWasmTypes } from "./aminotypes";
 export { CosmWasmClient } from "./cosmwasmclient";
-export { SigningCosmWasmClient, SigningCosmWasmClientOptions } from "./signingcosmwasmclient";
+export {
+  defaultGasLimits,
+  SigningCosmWasmClient,
+  SigningCosmWasmClientOptions,
+} from "./signingcosmwasmclient";